+++ backend/src/main/java/com/woowacourse/moamoa/review/service/ReviewService.java
@@ -1,12 +1,13 @@
 package com.woowacourse.moamoa.review.service;
 
-import com.woowacourse.moamoa.common.exception.UnauthorizedException;
 import com.woowacourse.moamoa.member.domain.Member;
 import com.woowacourse.moamoa.member.domain.repository.MemberRepository;
 import com.woowacourse.moamoa.member.service.exception.MemberNotFoundException;
 import com.woowacourse.moamoa.review.domain.Review;
 import com.woowacourse.moamoa.review.domain.exception.WritingReviewBadRequestException;
 import com.woowacourse.moamoa.review.domain.repository.ReviewRepository;
+import com.woowacourse.moamoa.review.service.exception.ReviewNotFoundException;
+import com.woowacourse.moamoa.review.service.exception.UnwrittenReviewException;
 import com.woowacourse.moamoa.review.service.request.WriteReviewRequest;
 import com.woowacourse.moamoa.study.domain.Study;
 import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
@@ -38,17 +39,14 @@ public class ReviewService {
         return reviewRepository.save(review).getId();
     }
 
-    public void deleteReview(final Long githubId, final Long studyId, final Long reviewId) {
+    public void deleteReview(final Long githubId, final Long reviewId) {
         final Member member = memberRepository.findByGithubId(githubId)
                 .orElseThrow(MemberNotFoundException::new);
-//        final Review review = reviewRepository.findByAssociatedStudyAndId(new AssociatedStudy(studyId), reviewId)
-//                .orElseThrow(ReviewNotFoundException::new);
-
-        final Review review = reviewRepository.findByStudyIdAndId(studyId, reviewId)
+        final Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(ReviewNotFoundException::new);
 
         if (!review.isReviewer(member.getId())) {
-            throw new UnauthorizedException("");
+            throw new UnwrittenReviewException();
         }
 
         reviewRepository.deleteById(reviewId);
