+++ backend/src/main/java/com/woowacourse/moamoa/review/service/exception/ReviewNotFoundException.java
@@ -1,8 +1,9 @@
-package com.woowacourse.moamoa.review.service;
+package com.woowacourse.moamoa.review.service.exception;
 
 import com.woowacourse.moamoa.common.exception.NotFoundException;
 
 public class ReviewNotFoundException extends NotFoundException {
+
     public ReviewNotFoundException() {
         super("후기를 찾을 수 없습니다.");
     }
